Alternative way to do cards
I added an alternate way to do the Cards. I used a class called Card and created an instance of the card in the opponent's deck. The opponent's deck now uses objects and the player's deck uses dicts. We are currently undecided on which method to use, so I did it this way to help us decide which method would be better to use during our week 2 meeting.