Added a deck and fixed some errors
I added a deck, called Steel Sun, and fixed some errors in the Class Card in the GameManager file. I have an error with the two attack damages that will need to be addressed. It was some kind of Key error. I also realized that when Chris reorganized everything he left out the functions in the gamestate file like the randomizing decks, the setup functions, and that stuff. I will work on getting them ported over to Chris' new GameManager class, or where ever they need to go now.