Merged GameManager with gamestate
I merged the functions from gamestate to the new GameManager class. I also made a copy of the deck to the opponent, so each player has a Steel Sun deck now. Tested the old functions in the new class and everything is working. I fixed the player's functions to work with the new class based approach.