End of the Night dump
This is just a upload of some small things I have done before I go to bed. I know Chris said he wanted to work on it tomorrow, so I wanted to assure he had the latest version. Not much substantial in this build. Most of the stuff I added was commented out because of errors. I will work on it with fresh eyes. I did get some win conditions added and fixed a few bugs. Started working on setting up board with multiple basics.