Cleaned and simplfied code
I cleaned up the code and simplified the actions. There was a lot of legacy code that was commented out that I removed. I simplified the attacks to only use basic attacks and removed everything but energy and attack from the game. We still have some issues with the AI recommending moves. Working on fixing the AI and the displaying of the game board to better see what is going on. I also added a logging feature.