Fixed and organized the game engine
I fixed the game engine so that all of the basic functions are working. There is still some issues with the MCTS that causes it to save moves when it should be doing it to the copy. I am working on further organizing the files into easier to follow files.